user specificity --added

+++ controllers/todoController.js
@@ -1,12 +1,17 @@
 // import the model Todo
 const Todo = require('../models/todo');
+const User = require('../models/user');
 
 const todoController = {
     getAllTodos: async (request, response) => {
         try {
-            const todos = await Todo.find({}, {});
+            // get the userId from the request object
+            const userId = request.userId;
 
-            response.status(200).json(todos);
+            // find all todos for the user
+            const todos = await User.findById(userId).populate('todos').select('todos -_id');
+
+            response.status(200).json(todos.todos);
         } catch (e) {
             return response.status(500).json({ message: 'Error fetching all todos...', error: e.message });
         }
@@ -14,6 +19,16 @@ const todoController = {
     getTodoById: async (request, response) => {
         try {
             const { id } = request.params;
+
+            // check if the todo belongs to the user
+            const userId = request.userId;
+
+            const user = await User.findById(userId);
+            const todoExists = user.todos.includes(id);
+
+            if (!todoExists) {
+                return response.status(403).json({ message: 'You are not authorized to read this todo' });
+            }
 
             const todo = await Todo.findById(id);
 
@@ -37,6 +52,18 @@ const todoController = {
             // save the data to the database
             const savedTodo = await newTodo.save();
 
+            // get the userId from the request object
+            const userId = request.userId;
+
+            // push the todo id to the user model
+            const user = await User.findById(userId);
+            user.todos.push(savedTodo._id);
+            await user.save();
+
+            // add the user id to the todo model
+            savedTodo.user = userId;
+            await savedTodo.save();
+
             // send a response to the frontend
             response.status(200).json({ message: 'New todo created successfully', data: savedTodo });
         } catch (e) {
@@ -46,6 +73,16 @@ const todoController = {
     updateTodo: async (request, response) => {
         try {
             const { id } = request.params;
+
+            // check if the todo belongs to the user
+            const userId = request.userId;
+
+            const user = await User.findById(userId);
+            const todoExists = user.todos.includes(id);
+
+            if (!todoExists) {
+                return response.status(403).json({ message: 'You are not authorized to update this todo' });
+            }
 
             const { title, description, isDone } = request.body;
 
@@ -61,8 +98,22 @@ const todoController = {
             // get the id from the request params
             const { id } = request.params;
 
-            // perform deletion
+            // check if the todo belongs to the user
+            const userId = request.userId;
+
+            const user = await User.findById(userId);
+            const todoExists = user.todos.includes(id);
+
+            if (!todoExists) {
+                return response.status(403).json({ message: 'You are not authorized to delete this todo' });
+            }
+
+            // perform deletion from todos collection
             const deletedTodo = await Todo.findByIdAndDelete(id);
+
+            // delete the todo id from the user model
+            user.todos.pull(id);
+            await user.save();
 
             if (!deletedTodo) {
                 // this means, deletedTodo is null
